Added I Term Decay. This bleeds off the accumulated I Term, when the locomotive is slowing to a full stop.
I Term Decay prevents Rollback Runaway (A dangerous condition where the
engineer has commanded the locomotive to stop from a high speed.
The locomotive briefly stops, but not long enough for the I Term to wind
back to zero, to allow the entire train to stop safely.
This situation induces Integral Windup as the control loop is completely
unaware that the locomotive is now traveling in the opposite direction
to what the engineer has commanded it to, & it is trying to do
everything in its power to stop the locomotive. Which, consequently,
causes the locomotive to increase speed, gather momentum & ultimately
derail) conditions from happening.

Traditionally, hitting the emergency brake early enough was used to nip
a rollback runaway in the bud.
This relied on the engineer to have reasonably fast reaction times, if
they were to save the entire train from disaster.
I Term Decay removes this requirement from the locomotive engineer.

Signed-off-by: Cassandra "ZZ Cat" Robinson <nicad.heli.flier@gmail.com>