I Term Relax part 1/2: Added 'pidSetItermRelax()' function. This is a placeholder function for when I get around to properly implementing I Term Relax.
The main way I want to do it is, I want it to partially bleed the I Term
off, when the locomotive is on flat/level grade of track; & when the
locomotive is climbing/descending a hill, that is when the full I Term
is used.

I can think of two more situations where I Term Relax can be used:
Yard work, where the locomotive is traveling at low speeds & is
regularly stopping & starting.
This is also partly the reason why I added I Term Decay.

Signed-off-by: Cassandra "ZZ Cat" Robinson <nicad.heli.flier@gmail.com>