Rewrote the entire Player Controls interface.
Player Controls now checks flags for whether-or-not a valid control
was updated.

Player Controls now have detailed fault reporting.

Latching Direction Fault.
When a Fault is detected, certain conditions must be met, before the
Direction Fault can be reset.
These are:
1.) The Emergency Brake must be released.
2.) Throttle must be set to Idle.
3.) Locomotive must be stopped.
4.) Direction must be set to Neutral.
Setting the Reverser to Neutral will clear this fault.

Latching Throttle Fault.
When this Fault is detected, these conditions must be met to reset it:
1.) Emergency Brake must be released.
2.) Locomotive must be stopped.
3.) The Reverser must be left in the same direction it was in, when the
Throttle Fault was triggered.
4.) Throttle must be set to Idle.
Setting the Throttle to Idle will clear this Fault.

All controls other than the one that caused a Fault will be locked out,
until the corresponding Fault is reset.

Emergency Brake interlocks with Direction, Throttle & Brake.
When the Emergency Brake is applied, the Direction, Throttle & Brake are
isolated (IE You cannot use them) until the Emergency Brake is released.

Signed-off-by: Cassandra "ZZ Cat" Robinson <nicad.heli.flier@gmail.com>