OTA Local: Completely overhauled the way the Version File is validated.
The Version File is now validated using a cyclic redundancy checksum on
the raw data, before it gets decoded from .json format.

Additionally, I now have a function that identifies oneself if they are
a developer of RailDriver (such as yours truly) or not.
This will enable me to automatically update the Version File, as I bump
RailDriver's semantic versioning up, over time.

I also plan to add on to my Developers' Whitelist in the Version File,
so my contributors have access to that same automatic update feature.

Signed-off-by: Cassandra "ZZ Cat" Robinson <nicad.heli.flier@gmail.com>