PIDF: Refactored Proportional Modes Proportional on Error & Proportional on Measurement are now switched dynamically between each other based on the accumulated I-Term & the I-Term's rate of change.
The higher the I-Term & the I-Term's rate of change is, the more the
control loop favors Proportional on Error.
The lower the I-Term & the lower the I-Term's rate of change is, the
more the control loop will favor Proportional on Measurement

This brings about the best of both worlds in terms of having the
control loop be fast acting to disturbances when needed, & the control
loop only maintaining a constant velocity when needed.

Signed-off-by: Cassandra "ZZ Cat" Robinson <nicad.heli.flier@gmail.com>